rename package
me.adamcraftmaster.schema
is now
me.adamcraftmaster.mapper
