+++ WeatherAPIcomLib/src/main/java/me/adamcraftmaster/ForecastLib.java
@@ -3,7 +3,7 @@ package me.adamcraftmaster;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import me.adamcraftmaster.exceptions.JSONGetException;
-import me.adamcraftmaster.schema.currentforecast.CurrentForecast;
+import me.adamcraftmaster.mapper.currentforecast.CurrentForecast;
 import me.adamcraftmaster.utils.JSONParserUtil;
 
 /**
