+++ WeatherAPIcomLib/src/main/java/me/adamcraftmaster/mapper/currentforecast/package-info.java
@@ -1,2 +1,2 @@
 /** This package contains JSON objects for mapping the forecast.json file */
-package me.adamcraftmaster.schema.currentforecast;
+package me.adamcraftmaster.mapper.currentforecast;
