+++ WeatherAPIcomLib/src/main/java/me/adamcraftmaster/mapper/currentweather/CurrentWeather.java
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package me.adamcraftmaster.schema.currentweather;
+package me.adamcraftmaster.mapper.currentweather;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
@@ -27,81 +27,57 @@ import java.util.Map;
 import javax.annotation.Generated;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({"text", "icon", "code"})
+@JsonPropertyOrder({"location", "current"})
 @Generated("jsonschema2pojo")
-public class Condition {
+public class CurrentWeather {
 
-  @JsonProperty("text")
-  private String text;
+  @JsonProperty("location")
+  private Location location;
 
-  @JsonProperty("icon")
-  private String icon;
-
-  @JsonProperty("code")
-  private int code;
+  @JsonProperty("current")
+  private Current current;
 
   @JsonIgnore private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-  /** @return String Weather condition text */
-  @JsonProperty("text")
-  public String getText() {
-    return text;
+  /** @return Location */
+  @JsonProperty("location")
+  public Location getLocation() {
+    return location;
   }
 
-  /** @param text Weather condition text */
-  @JsonProperty("text")
-  public void setText(String text) {
-    this.text = text;
+  /** @param location */
+  @JsonProperty("location")
+  public void setLocation(Location location) {
+    this.location = location;
   }
 
   /**
-   * @param text Weather condition text
-   * @return Condition
+   * @param location
+   * @return CurrentWeather
    */
-  public Condition withText(String text) {
-    this.text = text;
+  public CurrentWeather withLocation(Location location) {
+    this.location = location;
     return this;
   }
 
-  /** @return String Weather icon url */
-  @JsonProperty("icon")
-  public String getIcon() {
-    return icon;
+  /** @return Current */
+  @JsonProperty("current")
+  public Current getCurrent() {
+    return current;
   }
 
-  /** @param icon Weather icon url */
-  @JsonProperty("icon")
-  public void setIcon(String icon) {
-    this.icon = icon;
-  }
-
-  /**
-   * @param icon Weather icon url
-   * @return Condition
-   */
-  public Condition withIcon(String icon) {
-    this.icon = icon;
-    return this;
-  }
-
-  /** @return int Weather condition unique code */
-  @JsonProperty("code")
-  public int getCode() {
-    return code;
-  }
-
-  /** @param code Weather condition unique code */
-  @JsonProperty("code")
-  public void setCode(int code) {
-    this.code = code;
+  /** @param current */
+  @JsonProperty("current")
+  public void setCurrent(Current current) {
+    this.current = current;
   }
 
   /**
-   * @param code Weather condition unique code
-   * @return Condition
+   * @param current
+   * @return CurrentWeather
    */
-  public Condition withCode(int code) {
-    this.code = code;
+  public CurrentWeather withCurrent(Current current) {
+    this.current = current;
     return this;
   }
 
@@ -122,9 +98,9 @@ public class Condition {
   /**
    * @param name
    * @param value
-   * @return Condition
+   * @return CurrentWeather
    */
-  public Condition withAdditionalProperty(String name, Object value) {
+  public CurrentWeather withAdditionalProperty(String name, Object value) {
     this.additionalProperties.put(name, value);
     return this;
   }
