+++ WeatherAPIcomLib/src/main/java/me/adamcraftmaster/mapper/currentweather/package-info.java
@@ -1,2 +1,2 @@
 /** This package contains JSON objects for mapping the current.json file */
-package me.adamcraftmaster.schema.currentweather;
+package me.adamcraftmaster.mapper.currentweather;
